+++ packages/preact-query/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tanstack/preact-query
 
+## 5.91.1
+
+### Patch Changes
+
+- refactor(preact-query/useQueries): remove unreachable 'willFetch' branch in suspense promise collection ([#10188](https://github.com/TanStack/query/pull/10188))
+
 ## 5.91.0
 
 ### Minor Changes